requirements.txt
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
click==8.1.7
colorama==0.4.6
fastapi==0.115.5
Flask==3.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.10.1
pydantic_core==2.27.1
pytz==2024.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
Werkzeug==3.1.3
